chore: 연관 컴포넌트 한 파일 내에 정리, 절대 경로 연결 적용 안된 파일 수정, 기타 스타일 수정
