+++ src/Components/Modals/PaymentModal/OrderConfirm.tsx
@@ -1,12 +1,7 @@
-import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
 import { selectMenuListState } from "@/lib/state";
-import {
-  PaymentBox,
-  BtnGroup,
-  ConfirmButton
-} from "../../../Page/Client/ClientSelectList/styles";
+import { BtnGroup, ConfirmButton, PaymentBox } from "./styles";
 
 interface ConfirmProps {
   orderNumber: number;
