+++ src/Page/Client/ClientSelectList/ClientSelectListPage.tsx
@@ -1,17 +1,23 @@
 import { useState } from "react";
 import { useRecoilValue } from "recoil";
 import { useNavigate, useParams } from "react-router-dom";
-import { selectMenuListState } from "@/lib/state";
-import { calculateTotalAmount } from "@/lib/utils";
+import { AiFillMinusCircle } from "react-icons/ai";
+import { IoIosAddCircle } from "react-icons/io";
+import { Images, Modal, Noimage } from "@/Components";
 import { PaymentModal } from "@/Components/Modals/PaymentModal";
-import { Modal } from "@/Components";
-import { MenuListItem } from "./MenuListItem";
+import { IOrderSelectedItem } from "@/lib/state/productItemState";
+import { calculateTotalAmount, translateLocalCurrency } from "@/lib/utils";
+import { selectMenuListState } from "@/lib/state";
 import { OrderStateBar } from "../OrderStateBar";
 import {
   Header,
   MenuListWrapper,
   ResetButtonWrapper,
-  ResetButton
+  ResetButton,
+  MenuListItemBox,
+  MinusCountButton,
+  AddCountButton,
+  DeleteButton
 } from "./styles";
 import { useHandleSelectMenu } from "../hooks";
 
@@ -59,3 +65,58 @@ const ClientSelectListPage = () => {
 };
 
 export default ClientSelectListPage;
+
+interface MenuListItemProps {
+  item: IOrderSelectedItem;
+}
+
+const MenuListItem = ({ item }: MenuListItemProps) => {
+  return (
+    <MenuListItemBox>
+      {item.imageUrl ? (
+        <div style={{ padding: "0.5rem 1rem 0.5rem 0" }}>
+          <Images src={item.imageUrl} alt={item.name} />
+        </div>
+      ) : (
+        <Noimage />
+      )}
+      <MenuListItemDetail item={item} />
+      <RemoveMenuListItem item={item} />
+    </MenuListItemBox>
+  );
+};
+
+const MenuListItemDetail = ({ item }: MenuListItemProps) => {
+  const { handleMinusCount, handleAddCount } = useHandleSelectMenu();
+
+  return (
+    <div>
+      <h2>{item.name}</h2>
+      {item.option && <p>선택옵션: {item.option}</p>}
+      <p>
+        주문수량:
+        <MinusCountButton onClick={() => handleMinusCount(item)}>
+          <AiFillMinusCircle />
+        </MinusCountButton>
+        {item.totalCount}
+        <AddCountButton onClick={() => handleAddCount(item)}>
+          <IoIosAddCircle />
+        </AddCountButton>
+      </p>
+    </div>
+  );
+};
+
+const RemoveMenuListItem = ({ item }: MenuListItemProps) => {
+  const { handleDelete } = useHandleSelectMenu();
+
+  return (
+    <div>
+      <p className="price">
+        총 가격:&nbsp;
+        {translateLocalCurrency(item.totalPrice, "ko-KR")}원
+      </p>
+      <DeleteButton onClick={() => handleDelete(item)}>삭제하기</DeleteButton>
+    </div>
+  );
+};
