--- src/Page/Client/ClientSelectList/MenuListItem/index.ts
@@ -1,1 +1,0 @@
-export { default as MenuListItem } from "./MenuListItem";
